The User Manual
Here is the Word Document draft and PDF final if acceptable.  I went through the basics of using the program with some pictures too. 
If there is anything to be edited or added feel free to do so, but I also made a PDF already if there is nothing to be done here.